+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 2023/05/31: [#36](https://github.com/vestris/slack-invite/pull/36): Add a link to the Slack workspace on the invite page - [@accorvin](https://github.com/accorvin).
 * 2023/01/16: Upgraded to Ruby 2.7.7 - [@dblock](https://github.com/dblock).
 * 2022/06/12: Replaced Travis-CI with GHA - [@dblock](https://github.com/dblock).
 * 2020/04/23: Added `set sender` to change which Slack admin sends invitations - [@dblock](https://github.com/dblock).